requirements.txt
accelerate==1.10.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==1.4.4
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
auto-gptq==0.7.1
av==15.1.0
bitsandbytes==0.48.2
braceexpand==0.1.7
Brotli==1.0.9
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cmake==4.2.1
codecov==2.1.13
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coverage==7.4.4
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
Cython==3.0.10
datasets==4.5.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.0
distro==1.9.0
einops==0.7.0
einops-exts==0.0.4
evaluate==0.4.6
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.19.1
flake8==7.0.0
fonttools==4.60.2
frozenlist==1.4.1
fsspec==2025.10.0
gekko==1.3.2
git-lfs==1.6
gmpy2==2.1.2
gradio==3.44.4
gradio_client==0.5.1
groundingdino==0.1.0
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.36.1
idna==3.11
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.0.0
intel-cmplr-lib-ur==2025.3.2
intel-openmp==2025.3.2
interrogate==1.7.0
iopath==0.1.10
ipykernel==6.29.3
ipython==8.18.1
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.7
kwarray==0.6.18
lanms_neo==1.0.2
lazy_loader==0.4
lit==18.1.8
llava==0.0.1.dev0
lmdb==1.4.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mkl==2025.3.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mmcv==2.0.0
mmdet==3.0.0
mmengine==0.10.3
mmocr==1.0.0
model-index==0.1.11
modelscope==1.34.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.18
nest_asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==2.0.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onemkl-license==2025.3.1
openai==1.23.1
opencv-python==4.9.0.80
opencv-python-headless==4.13.0.90
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.3
ordered-set==4.1.0
orjson==3.10.1
oss2==2.17.0
packaging==26.0
pandas==2.3.3
parameterized==0.9.0
parso==0.8.4
peft==0.11.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.3
platformdirs==4.2.0
pluggy==1.4.0
portalocker==2.8.2
progressbar2==4.4.2
prompt-toolkit==3.0.42
propcache==0.4.1
protobuf==5.26.1
psutil==7.2.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==21.0.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.7.0
pydantic_core==2.18.1
pydicom==2.4.4
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
pytest==8.1.1
pytest-cov==5.0.0
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-utils==3.8.2
pytz==2023.4
PyYAML==6.0.3
pyzmq==25.1.2
qwen-vl-utils==0.0.14
rapidfuzz==3.8.1
referencing==0.34.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.4.2
rouge==1.0.1
rpds-py==0.18.0
ruff==0.4.0
safetensors==0.7.0
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==80.10.2
shapely==2.0.4
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
stack-data==0.6.2
starlette==0.37.2
supervision==0.19.0
sympy==1.14.0
tabulate==0.9.0
tbb==2022.3.1
tcmlib==1.4.1
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.4.0
tifffile==2024.4.18
timm==0.9.16
tokenizers==0.22.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.8.0
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.6
triton==3.4.0
typer==0.12.3
typer-slim==0.21.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ubelt==1.3.5
umf==1.0.3
urllib3==2.6.3
uvicorn==0.29.0
visual-genome==1.1.1
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
wget==3.2
wheel==0.41.2
xdoctest==1.1.3
xxhash==3.6.0
yapf==0.40.2
yarl==1.9.4
zipp==3.23.0